modbus_rest_gateway/app/requirements.txt
fastapi
uvicorn
pymodbus==2.5.3
python-dotenv
